azurerm-blob-storage/requirements.txt
azure-storage-blob==12.14.1
azure-identity==1.12.0
